+++ project/include/matrix.h
@@ -7,6 +7,9 @@
 #define INPUT_ERR "Non-exising matrixes"
 #define MULTIPLY_ERR "Matrix can't be multiply"
 #define DIMENTION_ERR "Diffrent nums or sizes"
+#define MEM_ERR "The end of the file wasn't reached"
+#define EOF_ERR "Can't allocate the memory "
+#define EXIST_ERR "The file doesn't exist"
 
 typedef struct Matrix {
 size_t num_rows;
