+++ project/src/creating_matrix.c
@@ -2,11 +2,6 @@
 #include <stdlib.h>
 #include "matrix.h"
 
-
-#include <stdio.h>
-#include <stdlib.h>
-#include "matrix.h"
-
 #define NUMBER_MATRIX_SIZE 2
 
 #define ALLOCATE_ERR "Can't allocate memory"
@@ -52,6 +47,46 @@ Matrix* create_matrix_from_file(const char* path_file) {
                 }
             }
         }
-        fclose(matrix_file);
+        if (EOF) {
+            fclose(matrix_file);
+        } else {
+            fprintf(stderr, EOF_ERR);
+        }
     return matrix;
+}
+
+Matrix* create_matrix(size_t rows, size_t cols) {
+    if (!rows || !cols)
+        return NULL;
+    Matrix* new_matrix = malloc(sizeof(Matrix));
+    if (!new_matrix) {
+        fprintf(stderr, MEM_ERR);
+        return NULL;
+    }
+    new_matrix -> num_rows = rows;
+    new_matrix -> num_cols = cols;
+    new_matrix -> value = calloc(rows * cols, sizeof(double));
+    if (!(new_matrix -> value)) {
+        free(new_matrix);
+        fprintf(stderr, MEM_ERR);
+        return NULL;
+    }
+    return new_matrix;
+}
+
+int check_for_exist(const Matrix *matrix) {
+    if (!matrix || !(matrix->value)) {
+        fprintf(stderr, EXIST_ERR);
+        return 1;
+    }
+    return 0;
+}
+
+void free_matrix(Matrix* matrix) {
+    if (matrix != NULL) {
+        if (matrix -> value != NULL) {
+            free(matrix -> value);
+        }
+    free(matrix);
+    }
 }
