+++ project/src/dop_functions.c
@@ -2,6 +2,53 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include "matrix.h"
+#include "dop.h"
+
+Matrix* adj(const Matrix* matrix) {
+    if (check_for_exist(matrix))
+        return NULL;
+    if ((matrix->num_rows != matrix->num_cols) || (matrix ->num_cols == 0)) {
+        return NULL;
+    }
+    size_t n = matrix->num_rows;
+    if (n < 1) {
+         puts("uncountable det");
+         return NULL;
+    }
+    Matrix* adject = create_matrix(n, n);
+    if (n == 1) {
+        adject -> value[0] = 1;
+        return adject;
+    }
+    double minor_det;
+    for (size_t i = 0; i < n; i++) {
+        for (size_t j = 0; j < n; j++) {
+            Matrix* minor = get_minor(matrix, j, i);
+            if ((minor != NULL) && !(det(minor, &minor_det))) {
+                int minor_sign = ((i + j) % 2)? -1: 1;
+                adject -> value[i * n + j] = minor_sign * minor_det;
+            }
+        free_matrix(minor);
+        }
+    }
+    return adject;
+}
+
+Matrix* inv(const Matrix* matrix) {
+    if (check_for_exist(matrix))
+        return NULL;
+    double determinant;
+    int buff = det(matrix, &determinant);
+    if ((determinant = 0 || buff)) {
+        puts("error determinant");
+        return NULL;
+    }
+    double value = 1/determinant;
+    Matrix *adjected = adj(matrix);
+    Matrix *invert_matr = mul_scalar(adjected, value);
+    free_matrix(adjected);
+    return invert_matr;
+}
 
 int det(const Matrix* matrix, double* val) {
     if (check_for_exist(matrix)) {
@@ -22,7 +69,7 @@ int det(const Matrix* matrix, double* val) {
         double determinant = 0;
         double minor_det;
         for (size_t i = 0; i < n; i++) {
-            Matrix* minor = delete_i_j(matrix, 0, i);
+            Matrix* minor = get_minor(matrix, 0, i);
             if ((minor != NULL) && !(det(minor, &minor_det))) {
             determinant += minor_sign * minor_det * matrix -> value[i];
             minor_sign *= -1;
@@ -34,7 +81,7 @@ int det(const Matrix* matrix, double* val) {
     return 0;
 }
 
-Matrix* delete_i_j(const Matrix* matrix, size_t row, size_t col) {
+Matrix*get_minor(const Matrix* matrix, size_t row, size_t col) {
     if (!matrix) {
     return NULL;
     }
